refactor(M): reorganize component imports and update Pagination
