+++ frontend/src/components/common/Stepper.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { cn } from '@/lib/utils';
 import { CheckIcon } from 'lucide-react';
-import { Button } from '@/components/ui/button';
+import { Button } from '@/components/ui/Button';
 import { StepperProps } from '@/interfaces/StepperProps';
 
 const Stepper: React.FC<StepperProps> = ({
