+++ frontend/src/components/layout/AboutHeader.tsx
@@ -1,5 +1,5 @@
 import Link from 'next/link';
-import { Button } from '@/components/ui/button';
+import { Button } from '@/components/ui/Button';
 import { ArrowRight } from 'lucide-react';
 import { useTranslations } from 'next-intl';
