+++ frontend/src/components/ui/Button.tsx
@@ -14,10 +14,13 @@ const buttonVariants = cva(
           'bg-transparent outline-1 outline-solid outline-bw-20 text-bw-70 hover:outline-2 hover:outline-bw-70',
         disabled: 'bg-bw-10 text-bw-40',
         destructive: 'bg-flame-50 text-white shadow-xs hover:bg-flame-50/80',
+        ghost: 'hover:bg-bw-10',
       },
       size: {
         default: 'rounded-md px-4 py-3',
         full: 'rounded-md px-4 py-3 w-full justify-center flex-1',
+        pagination: 'h-9 px-4 py-2 has-[>svg]:px-3 hover:bg-bw-10 hover:rounded-md',
+        paginationIcon: 'size-9 rounded-md ',
       },
     },
     defaultVariants: {
