+++ frontend/src/components/ui/Slider.tsx
